[Added] JTL Prelim Implementation [Updated] Packet wiki updated with baseline structs ToDo: Handle at the SWG protocol level.
This is a premature version of the overall JTL scheme in Core3. We currently
do not send the client any of this info, yet. The code will be matured
and slowly move to an actual ingame feature. This was committed early to prevent
excessive amounts of conflicts in the future, especially with the survey/resource
revision coming up.


git-svn-id: http://svn2.assembla.com/svn/swgemu/core3devel/MMOCoreORB@105 c3d1530f-68f5-4bd0-87dc-8ef779617e40