The commands @kick @kill
have been changed.

The format now is
@command <name>

If you supply a name, it is always first priority, regardless what your toon has currently targeted.

If you supply no name, your current target will be choosen for the action.

The <name> option has no rangelimit (of course) and is working on always, no matter
on which planet the target-toon is.

The commands will now send a message to the player so he is aware that administrative action was used on him.


========================================
New command:
@mutePlayer

This command is following the above rules and is muting the spatial chat for a single player.
This command will send a message to the player so he is aware that administrative action was used on him.

@mutePlayer works as a toggle - re-apply the command and the player is able to use the (spatial) chat again.

git-svn-id: http://svn2.assembla.com/svn/swgemu/core3devel/MMOCoreORB@402 c3d1530f-68f5-4bd0-87dc-8ef779617e40