[Added] Waypoint Usage
MAKE SURE YOU UPDATE THE WAYPOINTS TABLE WITH THE /sql/update_waypoints.sql

this is the waypoint patch, so core3 will load and safe waypoints to the database.
Ingame, waypoints are handled by a sorted vector (waypointList).


Also there are functions now saving

• Friendlist (not fully in yet)
• Ignorelist (not fully in yet)
• Waypoints (not fully in yet)

if the client is crashing, kicked, booted, logging out, whatever.



In the ChatManagerImplementation.CPP is a commented block for the command "@saveToon".

This command is NOT under admin rights and can be used to save the above lists on demand (testing etc.)


git-svn-id: http://svn2.assembla.com/svn/swgemu/core3devel/MMOCoreORB@421 c3d1530f-68f5-4bd0-87dc-8ef779617e40