[Added] Completed Friendslist/findFriend [Updated] sql/swgemu.sql with the new update_friendlist.sql/update_waypoints.sql [Updated] sql/update_waypoints.sql [Updated] sql/update_friendlist.sql
This is the first patch addressing community functions:

1.	addfriend <name>
2.	removefriend <name>
3.	findfriend <name>


Findfriend will create a waypoint to the friend (if you are on his friendlist),
therefore a new field has been added to the waypoint object:

InternalNote

This field can be used for any “flagging”. Sometimes the server needs to know if a 
waypoint has to be updated or generated new. This WP cant be identified by its name.

Imagine /findfriend bob – the server will generate a WP named bob. If I do an update on bob (/findfriend bob again) the WP needs to be updated. 
The server cant identify which WP to update since I could also have another waypoint
named “bob” which I generated earlier, pointing to bob’s house.

I’ve added the “internalNote” member to the WaypointObject, there we can store such needed indentifier-flags. In case of findfriend, this field holds the value “FINDFRIEND:BOB”.

Now the server can identify this WP and upate – or generate a new one if there is no such waypoint yet.

The field could be also used for mission-waypoint etc. It’s a string field so you can code your logic for that field like you want, eg. “MISSION:ID463-STAGE1-WP2”


Addfriend will check if the player is already friend or not. Addfriend will also make sure, the to-add-player is existing in the database.

There are certain places within the client, where a character name will be converted (first letter uppercase) so the friend-functions are converting the name (only) internally to lowercase to make sure the wanted function will operate flawless.

There are some changes/additions in “notifySceneReady” since the client is tweaked to multi-zoning server (SoE). The client will reset some internal values, counters and pointers, so “notifySceneReady” is used to fire some events after zoning a player.

NotifySceneReady has also been changed to show “Welcome to EMU” only for the first time logging in. Make sure you do not change the condition checks there (if online==LOGGINGIN) otherwise Findfriend (and later ignorefriend) will not work correctly after zoning!!


Friendlist is saving to the database and loading from there on login, so make sure, the attached 
mySQL-script is applied !

Findfriend “should” work without problems, also informing your friend about login, logoff etc.

Please report any bug to the forums or PM me:
Movealong (Forum Nick)
Farmer_John (IRC)


Credits and a BIG thank-you goes to KYLE !
He was the one doing the packetwork for this functions after I gave up on them after a week…

THANK YOU KYLE!

We found a very nasty “unknown integer” in the packets which was a tough one to “de-mystify”.
Of course, its “de-mystified” finally, it’s called in the code “magicnumber” (friendsMagicNumber, magicNumber).

This variable is UBER-ESSENTIAL to make findfriend work, so don’t even touch it plz…


Have fun !
Farmer_John

git-svn-id: http://svn2.assembla.com/svn/swgemu/core3devel/MMOCoreORB@434 c3d1530f-68f5-4bd0-87dc-8ef779617e40