Update patch set 1
Patch Set 1:

The current structure is difficult for a human to read. The best example is try to find all of one tribe in Endor. The naming conventions are not consistent. Sometimes the creature has the verb in front of the specie and sometimes the specie is not even in the title at all. The alphabetical list it makes subspecies scattered. Renaming files consistently like specie_variation breaks things if they have been referenced to spawn anywhere. Moving stuff to a folder just needs the server_object file updated so its safer to not break the code. The mobile folder has been unfinished since the last big reconstruct. The /all folder is the remaining creatures that didn't get organized which is the last couple letters of the alphabet. The /misc folder has lots of duplicates. The dungeon mobiles were scattered between the dungeon folder, planet folder, and misc folder. Bio engineered creatures just had a _be sufix and were organized by planet which made them really hard to round up if something needed to be changed. The organization helps find things much quicker and it makes mistakes more obvious. I can open the specie folder and see if all the variations are present. Using a find function won't show file names without the specie in the title. Several spelling mistakes were corrected during the combing process. Species on the wrong planet and duplicates have also been found. The swg project has been coming along nicely and I felt there should be more polish for the final product that will be distributed. I mostly work on the mobiles folder and have been annoyed with the current structure for the past year which is why I put so much work into it. I want the final product to not have a learning curve like having to use boba_fett in the /tatooine folder because the boba_fett in the /misc folder is disabled and doesn't converse.

Patch-set: 1