Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

I'm going to shelve this one for a bit until I get http://76.73.75.106:8080/#/c/96/ done. It needs a bit of tweaking and I should be able to use the method here that Loshult asked me to make in the other one.

Patch-set: 2