Update patch set 3
Patch Set 3:

for the loot groups I made a group for each of the armor types and another group that included all the armor pieces.

As far as the stats on the armor goes, I used the experimentation values from the object/tangible/wearable/armor/... i set all the encumberances and armor integrity to the experimental values for each of the pieces. 

as far as the effectiveness goes, I set base effectiveness to: min = 4, max = (experimental value -1). some of them had experimental minimums of 1 while others had 5. special protections I set to: min = 10, max = experimental value. (set min of 10 so the special protection will always be higher then the base effectiveness). the only armor type that is different from this is composit, and that is because it didn't have an experimental value listed for special protections. i set composits special protection to 10 - 50 so that its special will always be higher then its base effectiveness (composits base effectiveness is set to 4-39)

Patch-set: 3