Update patch set 2
Patch Set 2:

Loshult, I did as you instructed and created a method in PlanetManager to find the closest world floor.

I went ahead and rolled the other patch I had up to set the correct height for dismounting into this one as it is doing the same thing and could use the same method.

I searched the code as you suggested. The only other place that CollisionManager::getWorldFloorCollisions() was being called was in PlayerManager::updateSwimmingState(). As this isn't using the intersections for the same purpose, I left it alone.

I believe that the "position verification methods" you mention in your comment is likely referring to the DataTransformWithParentCallback class. It does do a check inside it's run() method that is very similar. It is, however, calling the CollisionManager::getCellFloorCollision() method to get it's collision points. It seemed different enough that I didn't want to try and merge it into the method I made without further direction.

Patch-set: 2