Update patch set 2
Patch Set 2:

After doing a bit more digging, I see that DataTransformWithParent is for position verification inside non-world cells and there's a DataTransform class for open world verification. Looking through the class it does not look like it's doing any verification that the Z coordinate is a valid world floor (and I'm guessing is why players are sometimes getting stuck in the air). It probably should so I think I will add a call to my new method and try it out a bit. If it looks good I will recommit a new patch set with it included.

Patch-set: 2