Update patch set 1
Patch Set 1:

That should be correct, but I would still appreciate some further documentation on how to easily add new/modified files to index, and how to properly pull to HEAD/merge. Currently with many files added/modified in a large patch set such as this, it's kind of confusing to do via Git.

Patch-set: 1