Update patch set 1
Patch Set 1:

there was a long back and forth done on this topic in the architect forums http://www.swgemu.com/forums/showthread.php?t=67814 the experimentation on the extraction rate on these harvesters was acknowledged as a bug by soe dev on 12-31-04, then they acknowledged it was being worked on on 1-15-05. however the fix for this was not on the live servers until sometime between july 29th 2005 and 0ctober 29th 2005.

Patch-set: 1