Update patch set 5
Patch Set 5:

armor dropped as loot during pre-cu, a quick search on google images i found 4 screenshots of looted pre-cu exceptional and legendary armor. posted links to them here http://www.swgemu.com/forums/showthread.php?t=82858

what I'm not sure on is the values that should be set for the min and max effectiveness on the armor. which is why i used the experimental values. if the effectiveness on the armor needs to be adjusted i can do that.

Patch-set: 5