Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Again, please do not merge this yet, I wanted to do a initial push of my work so far for Loshult to more easily review.

In short, the current method of the rebel cell permission is incorrect as it was 14.1, as there are 'warning' STFs in the files that are used and you are usually warped to a previous cell, still need to figure out the logic behind that.

Patch-set: 1