Update patch set 1
Patch Set 1:

Info on drop locations came from:

http://www.swgemu.com/archive/scrapbookv51/data/20070131141609/index.html

and

http://www.mighty-gamers.net/forum/index.php?page=Thread&threadID=104

I based the drop chance on how many of each part/schematic is needed to make a full set of each recycler. I.e. a part that you only need 1 of will be rarer than one you need 8 of. I set them to roughly 1% overall per item needed.

In every case, I kept the overall loot chance and the chance for non-junk items roughly the same as previous. I simply lowered the chance of junk to add the chance for the new items.

Patch-set: 1