Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

It was brought to my attention that with my previous update to camp xp that someone could gets lots of xp by dropping and abandoning camps over and over in a short duration if many people were present. This is a change to curb that ability.

Patch-set: 1