Update patch set 1
Patch Set 1:

Note about this fix. The only way i was able to get the terminals to appear was to delete my clientobjects db. If accepted same will have to be done for nova and then bas.. Unless there is another way to fix this that i'm not aware of.

Patch-set: 1