Update patch set 1
Patch Set 1:

This is one of those which I think is way past the "grey" area things.

This particular issue was verifiably confirmed by the SOE devs in 2004. It was publicized as "being fixed" before the CU was even developed. 

The actual fix didn't get finished and implemented until during the CU. 

This one isn't a case of not being fixed because it wasn't consistent with earlier game mechanics or questionable as to what it was. 

The devs themselves agreed that there was a problem, agreed what the problem was, agreed to fix the problem, had it in the queue to be fixed and finally implemented that fix. Of all those steps, ONLY the implementation happened post-14.1. In fact, everything else happened and was documented publically by the devs prior to or during publish 11. It could (and should) have been implemented with publish 12, 13, or 14.

Most of my research can be found in the thread Mars' linked. I can provide all of it and more if necessary.

Patch-set: 1