Update patch set 1
Patch Set 1:

I decided that I needed to invert the logic for city structure registration in order to prevent commercial structures that are no longer in a city from being registered on the planetary map.

Before this patch the zone registered all objects and them the city, if not registered, unregistered it's civic structures.

After this patch the zone skips civic or commercial structures when registering objects but then the city, if registered, will register it's civic and commercial structures.

Patch-set: 1