Update patch set 1
Patch Set 1:

I got looking and there were not no-build zones for POIs as I had thought. Building was being restricted by a call to isInRangeWithPoi() inside PlanetManagerImplementation::isBuildingPermittedAt()

It was preventing building within 512m of any POI. I didn't like the arbitrary distance and felt that it was too large in many cases. I got with Scurby and together we decided to lower that to 150m and then add no-build zones to POIs where the 150m wasn't sufficient. We visited each to decide upon where to add zones.

Patch-set: 1