Update patch set 1
Patch Set 1:

It is not currently possible to disable the use of the string for one mission's waypoints. Either the whole series uses names from the string files or none of them do. Because this mission was missing one, I set it to use the default waypoint names. 

I added the custom names for this quest because the default waypoint names pull from that field.

Patch-set: 1