Update patch set 1
Patch Set 1:

The deleted item was a duplicate of the projectile_pistol_barrel_quest.lua which is included in the jabba_theme_park_reelo group.

The client name for the shellfish harvester schematic is "Shellfish Harvesting Tool" which doesn't indicate that it is a schematic. I renamed it to be consistent with other reward schematics.

Patch-set: 1