Update patch set 1
Patch Set 1:

Many quests/themeparks are supposed to only be available to those of a specific faction. The themepark logic does not currently support the ability to prevent those of the wrong faction from accepting and attempting the quest. The intent of this patch is to add in that functionality.

Patch-set: 1