Update patch set 1
Patch Set 1:

I don't see how I removed any functionality. Each task or themepark can have the required faction set for it individually. I don't know of any situation where we'd want one task or themepark to require the player to be in one faction for one mission then in a different for the next.

The only existing functionality I changed was that I made it so that factional permission maps no longer have to have the faction assigned for each, but will inherit the faction of the themepark/task that they are a part of.

Patch-set: 1