Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

(3 inline comments)

The armor not generating with full condition has to do with the manner in which themepark loot is generated. It's creating the items in the same manner that loot is created. At some point we can look into changing this behavior, but I don't think it's a big deal and therefore not a priority at this point.

Patch-set: 1