Update patch set 1
Patch Set 1:

As world spawn areas, the newbie areas were valid areas for mission terminals to pull lair spawns from.

This made it such that the lower level creatures on starter planets were much more likely to show up on the mission terminal than the higher level ones that were not in the starter list.

Patch-set: 1