Update patch set 2
Patch Set 2:

I'm not sure that I should create the utils folder. I can easily move gtest to another folder if anyone has a better placement of it.

I will later also add gmock to be able to create mocks for our classes to be used in the unit tests.

Patch-set: 2