Update patch set 1
Patch Set 1:

This is a precursor to property tax.  This will allow us to use the same functions that are used to collect normal maintenance.  Normal maintenance can be updated and fractions/decimals removed from the maintenance.  Property tax will be a percentage of the amount of normal maintenance paid.

Patch-set: 1