Update patch set 1
Patch Set 1:

Wasn't sure the best way to fix this.

The problem is that Commercial Structures, Civic Structures, and the City Hall are deserialized before the cityRank parameter.  So, notifyEnter is called before the cityRank is set and we think it is a clientRegion and don't add it to the completeStructureList.

Some of the other options could be.

1.  Iterate through the commercialStructureList, and civicStructureList when doing the citizenCleanup().

2.  Make commercialStructureList and civicStructureList transient so the strucutres are deserialized.

3.  Come up with a way to bypss the isClientRegion() check in notifyEnter for commerical and civic structures.

Patch-set: 1
Reviewer: Gerrit User 4 <4@87af08d3-5c3e-4272-a6b6-bb0c36cf31f2>
Label: Verified=0