Update patch set 1
Patch Set 1:

Looks like the CityRegion change to fix the cantina might have caused duplicate items to appear in the structures commercial structures and decorations vectors.

Not sure how that happened but it didn't happen to me locally.  Figured i'd offer up a quick way to fix it for now until we figure out how duplicates appeared on the list.

Also, tried to roll back to the previous version of CityRegion, plant structures and the decorations were the only ones to duplicate.

Not sure if this is a better option to rolling back the database and trying to figure out how it happened.

Patch-set: 1