Update patch set 1
Patch Set 1:

just a theory and not sure if this is possible or not but it is behaving like when readObject() is called, the objects are loaded into the area and added to the vector before it is deserialized and put into the vectors.  

it seems like the object zones and is added to the vector then the existing item in the db is deserialized and added onto it bypassing the no duplicate.

Patch-set: 1
Reviewer: Gerrit User 4 <4@87af08d3-5c3e-4272-a6b6-bb0c36cf31f2>
Label: Verified=0