Update patch set 3
Patch Set 3:

I'm not sure I exactly understand what you are asking.

That being said, I had considered listing the cells and ejection point in the dungeon screenplays and handling the ejection there. I opted for the method I used because it was easy to make a list of all of the dungeon cells on the planet, and to check if the player was in a dungeon cell.

Furthermore, I figured that this method was extensible. Any static cell would be able to easily have an ejection point specified for it. We could move people out of the Bestine ID salon as they log out, for example.

I'm certainly open to better/other ideas.

Patch-set: 3
Reviewer: Gerrit User 9 <9@87af08d3-5c3e-4272-a6b6-bb0c36cf31f2>
Label: Verified=0