Update patch set 1
Patch Set 1: Code-Review-1

Pushed this early for review.

This is what I'm thinking of for managing droid modules.  The way they are passed over in CraftingValues is pretty obscure.  Any feedback welcome.

The Vector of modules in DroidDeed and DroidObject are also not being persisted into the db.  I need some help with that.

Patch-set: 1
Label: Code-Review=-1