Update patch set 1
Patch Set 1:

I'll give it a shot.  I know this compiles and runs fine but the automated testing is based on old methods/calls.  I'll review and try pushing with the same change-id or just suck it up and start pushing with Eclipse from my dev machine.  Thanks for your feedback.

Patch-set: 1