Update patch set 1
Patch Set 1: Code-Review-1

The changes loos fine, however, not everything mentioned in the commit message appears to actually be implemented in this commit.

Specifically, it says it adds an LOS check to both commands but I only see it added to transfer force.

Also, I do not see where you're preventing a transfer of force to a player at max force.

Patch-set: 1
Label: Code-Review=-1