Update patch set 1
Patch Set 1:

Don't fix this in the individual commands. Fix it in the isXXX() checks in WEAO itself.

As a note: The server reads the client object types from the client files, so that is why this is a bug in the first place -- you can't fix it by adding a new SceneObjectType, but you can fix it by &&'ing in a jedi weapon check in WeaponObject.idl. This will also fix other bugs (like 5232) and is infinitely more elegant. You're looking at a 3-line change solution in a single file.

Addendum: You don't need a NULL check for weapon. It is guaranteed not to be NULL.

Patch-set: 1