Update patch set 1
Patch Set 1:

Oldest revision on swg.wiki.com states that questions were asked in random order, the answers were listed randomly, and that if you missed an answer you could try again. While one could interpret that as meaning you'd have to answer every question over again, that interpretation is obtuse. Were that the case, it would be clearly noted, as having to answer all the same questions over again is a nuisance that would not go unmentioned. This is also supported by the documentation in the scrap book. 

You can read all the patch notes for the entire and you will not find that SOE ever changed this quest after patch 14.1, so any documentation related to will suffice.

http://swg.wikia.com/wiki/Trivial_Librarian?diff=cur&oldid=27614

http://sassyswg.proboards.com/thread/158/naboo-librarian

https://www.google.ca/url?sa=t&rct=j&q=&esrc=s&source=web&cd=2&cad=rja&uact=8&ved=0CCQQFjAB&url=http%3A%2F%2Fwww.swgemu.com%2Farchive%2Fscrapbookv51%2Fdata%2F20070206110440%2Fquests_245.html&ei=I6a5U7naIcOTqAbhk4KIBQ&usg=AFQjCNHrc5B23XhVpxvxEsdNHa4Cs1HF9w&sig2=pS3eQGodg546roo23g_1tQ

Posting this was a waste of my time. I'm not going to do it again.

Patch-set: 1
Reviewer: Gerrit User 984 <984@87af08d3-5c3e-4272-a6b6-bb0c36cf31f2>
Label: Verified=0