Update patch set 1
Patch Set 1: Code-Review-1

I like the infrastructure you add here, but there are 2 problems with the code as it is. First, the waypoint needs to update as group players move around (this only updates them when missions get added/removed or when players get added/removed). Which leads to the second problem: this is far to CPU expensive for a task that does not need to be very precise -- the way to implement this needs to be thought about very carefully and a balance needs to be struck between CPU cycles and accuracy (probably best to do a semi-long time task in combination with roughing the distance comparisons significantly).

Patch-set: 1
Label: Code-Review=-1