Update patch set 2
Patch Set 2:

That is precisely what I think. You are performing squaring operations for every mission in the group every time it updates. That needs to be abated a bit, and performing the task once every few seconds (I think 30-60s) would really help with that in cases where it would be a problem. I'll look at the rest of the code when I get around to it (tomorrow) -- you did use the triangle inequality for distance checks, no? =D

Patch-set: 2
Label: Verified=0