Update patch set 3
Patch Set 3: Code-Review-1

(12 comments)

I like this much better (though there are a couple of things to fix still). I would also consider storing the task in GroupObject so that you can either cancel it and reschedule if you want to or simply not schedule it if one is already running. The way it is now (if you replace the ->run() with ->schedule() like you need to) it will schedule multiple tasks and not have any advantage.

I personally think you should cancel (Task::cancel()) and resubmit (Task::schedule()) so it waits until everyone is done changing everything up before it updates, but I'm willing to accept the other way if you think it's better.

If you want to talk about it, I am on IRC all day long (as dannuic, just PM me)

Patch-set: 3
Label: Code-Review=-1