Update patch set 1
Patch Set 1: Code-Review+1

The change explicitly states that hopper must have 1 or more full units for it to need emptying. 

Without this change, any partial value between 0 and 1 is considered as "not empty" by the engine, while the UI in the game only shows 0, which is not intuitive or reasonable for the end user.

This is caused by the value being a float rather than an int.

Patch-set: 1