Update patch set 1
Patch Set 1: Code-Review+1

Nope - I made that patch and it doesn't work because the data being evaluated was incorrect. This new patch does work because the data being evaluated is correct now. 

I wrote those very words here and in the mantis ticket and commented it in the code itself.

Patch-set: 1