Update patch set 1
Patch Set 1:

check http://review.swgemu.com:8080/#/c/2246/ again. It is not your code that was merged, but mine. It fixes it so that you don't even display 0 quantity resource containers in the hopper list unless the hopper is running.

Make sure you have actually tested again and that this is actually still an issue. And if it is, this is not the way to fix it, but the hopper needs to never show resource containers that are empty when it is turned off.

Patch-set: 1