Update patch set 1
Patch Set 1: -Code-Review

I did test it on the latest updates and that is how I discovered that the issue still occurs. As I described, the value being evaluated in your version, resourceHopper.size(), doesn't work because that value gets updated to 1 after any UI interaction is made to cause the harvester update function to be called. I traced this issue with logger output and it was clear as day. Therefore, that value is not an appropriate or even functional value to use for this purpose.

The way I have done it in this patch, basing the check on whether or not the hopper has any resources that need to be emptied (which is logical) works and will solve this annoyance for people. 

Furthermore, I don't see any issue with evaluating the correct data to come to the correct conclusion. The data is able to be used in this context and it should be used in this context to ensure the device functions as expected.

Patch-set: 1