Update patch set 1
Patch Set 1: Code-Review+1

Note that resourceHopper.size() is being changed to 1 or 2 even when the harvester is off, after any ui interaction, which is why  if(harvestAmount > 0 || !isOperating()) { in your patch also does not resolve the problem.

Test it yourself. I spent a couple hours today doing so...

Patch-set: 1
Label: Code-Review=+1