Update patch set 1
Patch Set 1:

Also, the slot configurations are all on the radial.  You click Configure and a subradial shows all the slots.  I'm not 100% sure if it did that or if it popped up a SUI to show the configuration and allow them to select a slot to configure.

This had the best description:
http://www.sunfall.eu/forums/viewtopic.php?f=1&t=1431

QA, let me know if you think slots should be a SUI and not on the radial.

Patch-set: 1
Label: Verified=0