Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

I'm not 100% sure that peace will trigger correctly with the new logic in CombatManager, and it triggered some discussion with the devs on how peace is actually supposed to work. Specifically, when is the peace icon supposed to show up? If you set the peace state, you will get the peace icon in the client, so it would be very useful if you could find out how this is supposed to function from live references.

The -1 is for the fact that we don't want to drop all commands, just combat commands.

Patch-set: 2
Reviewer: Gerrit User 1005 <1005@87af08d3-5c3e-4272-a6b6-bb0c36cf31f2>
Label: Code-Review=-1