Create patch set 2
Uploaded patch set 2.

Patch-set: 2
Subject: [Fixed] a mail is now sent to the mayor if a city's decoration, terminal, or trainer is destroyed due to lack of maintenance [Fixed] if a city cannot afford to pay the cost of it's specialization during it's weekly update, the specialization is removed [Fixed] if a city cannot afford to pay the cost of registration during it's update, the city is unregistered [Fixed] when a civic structure is destroyed due to lack of maintenance, only a structure destroyed email is now sent to the mayor. A structure decayed email is no longer also sent
Commit: 473be9e57546651b49893d1dc0e3729a140c871a