Create patch set 2
Uploaded patch set 2.

Patch-set: 2
Subject: Modify GetGCWDiscount() : the function was only able to set the GCW Discount as a penality. (only negative Values from the Lua) Add two new Values looserBonus and winnerBonus. Both can have positive or negative values to give either a Bonus or a Penalty The setting here was that the looserside get a panalty of 30% We set it on our server as a winnerBonus 30 and a looserbonus 0 Iam not sure but i think that was on pre cu equal. But i set the config you have had as a looserpenalty of 30%
Commit: 481278fca3cdd678c0408c9a2edf9144c195432b