Create patch set 3
Patch Set 3: Published edit on patch set 2

Patch-set: 3
Subject: Modify GetGCWDiscount() : the function was only able to set the GCW Discount as a penality. (only negative Values from the Lua) Add two new Values looserBonus and winnerBonus. Both can have positive or negative values to give either a Bonus or a Penalty The setting here was that the looserside get a panalty of 30% We set it on our server as a winnerBonus 30 and a looserbonus 0 Iam not sure but i think that was on pre cu equal. But i set the config you have had as a looserpenalty of 30%
Commit: e1a47765bc646db667c5677db377dbbe286333af