Create patch set 4
Patch Set 4: Published edit on patch set 3

Patch-set: 4
Subject: Modify GetGCWDiscount() : the function was only able to set the GCW Discount as a penality. (only negative Values from the Lua) Add two new Values looserBonus and winnerBonus. Both can have positive or negative values to give either a Bonus or a Penalty The setting here was that the looserside get a panalty of 30% We set it on our server as a winnerBonus 30 and a looserbonus 0 Iam not sure but i think that was on pre cu equal. But i set the config you have had as a looserpenalty of 30%
Commit: e8f0f1e713d38769d4c6d0ed4dca3fdb1012069e