Update patch set 9
Patch Set 9: Code-Review-2

As I looked over this, I decided that I really don't like the design of it. I think you should move it all into lua.

You're adding tables in lua with the xp types and ratios. I don't really see a need to load them into cpp. The only thing you use them for in cpp is to compare them to the player's xp list to compile a list of types to display in the sui, right? It would be much cleaner to just do it in lua. Oh, and storing a vector of strings on the playo that you recompile every time you get it is just a waste.

Another plus to doing this would also keep the village code isolated. A server that uses a different jedi progression system wouldn't need to have the fs experience types and ratios loaded.

If you need help or suggestions, feel free to ask.

Oh, and don't change the indentation style of the lua files.

Patch-set: 9
Reviewer: Gerrit User 9 <9@87af08d3-5c3e-4272-a6b6-bb0c36cf31f2>
Label: Code-Review=-2