[changed] Jedi heals unified
All Jedi healing abilities can be configured from their lua files.
This allows for much easier and safer (as in less buggy) ways to change
the behaviour of the heals. Previous bugs (potentially not reported)
of people healing without force, healing without animations due to not
being permitted etc. should be fixed with this change.

Change-Id: I0606e86e6dcc89849e0395348fc34567dd80fed3