[WIP] Jedi heals unified
I got somewhat upset about the state of all the jedi healing abilities
so I decided to refactor that stuff a bit...

The idea is to have everything configured from lua files on startup.
The difference between the abilites seem only to be the
attributes/states/wounds they are healing and the costs so unsing/
improving the ForceHealQueueCommand class seemed like a good thing.

I am aware that this code might be slower then the others.. I tried
to make it so one can reuse bits and pieces from sub-classe if so desired.

ToDo:
 - allow heals to be applicable to both a remote target and the healer
   (usef for poison, bleeding, disease,..)
 - minimal costs for force heals (needs some research)

Change-Id: I0606e86e6dcc89849e0395348fc34567dd80fed3