Update patch set 1
Patch Set 1:

The way I would interpret the patch note is not that you can take a sample while mounted, but merely that the command would function while mounted. Instead of giving an error when using /sample while mounted, you would automatically dismount, kneel, and then take a sample. It's certainly not verbose. Some additional evidence of correct behavior would be nice if it could be found.

Patch-set: 1
Label: Verified=0