Update patch set 2
Patch Set 2: Code-Review-1

(11 comments)

Things to consider: AI reactions (after pub 11, this is non-trivial, check the patch notes), most of the state checking is done automatically because we pull the states from the client files and check against that.

Things missing: There was a damage modifier while FD (check the patch notes), pub 13 made it so you can be incapped while FD, there should be a 5 second delay to perform ANY action (including stand), you need to add any and all spam/system messages (check the client, there are 2 spam messages and 3 system messages), check to make sure you are in combat before you can FD

Also, the use of the double buff is a bit hard to follow, I'd use an observer(DAMAGERECEIVED) to achieve the result you are looking for and simplifying the logic.

PS, I know a lot of this we discussed in IRC, but I wanted to leave notes for you in a more permanent place.

Patch-set: 2
Reviewer: Gerrit User 1525 <1525@87af08d3-5c3e-4272-a6b6-bb0c36cf31f2>
Label: Code-Review=-1