Update patch set 3
Patch Set 3: Code-Review-1

This is the wrong place for that config option. It's configured in visibility manager, but then only used in mission manager. Just configure it in mission manager.

Also, MissionManagerImplementation::getRandomPlayerBounty() needs to be redone to not be so goofy, but that can probably wait.

Patch-set: 3
Label: Code-Review=-1