Update patch set 10
Patch Set 10: Code-Review-1

(11 comments)

I would use a buff on the target that reactivates itself. In the activations, it either calls the choke command or sends the equivalent through combatmanager (look at powerboost for an example of how to do that). Make sure you can overwrite this buff. Remove the task.

Patch-set: 10
Label: Code-Review=-1