Update patch set 1
Patch Set 1:

The bug report doesn't say anything about how it should be displayed other than it not being in scientific notation. If you print it as "4.99 Million" Then you lose a TON of precision that is required when deciding the loadout of the ship (components had masses in the ~10k range). Further, I'm about 99% certain that all masses had a decimal precision to the tenths (given the one screenshot we have of ship mass in the UI), so mass should be appearing as "4999840.0". I asked Nee to post the screenshot I'm referring to in that bug, but you can see how snprintf() would be used here (or in our case, please use String::format()).

Patch-set: 1