Update patch set 1
Patch Set 1:

I implemented the String::format, using 7.1f, which is consistent with the screenshot that Nee posted in the bug comment. My only concern is that this may cause other things to display incorrectly, since this is being applied to the TangibleObject base. 
Is this an acceptable solution?

Patch-set: 1