Update patch set 1
Patch Set 1:

You're right about it affecting other things. I think this is a case where the type of the original variable determines how it was printed, which could be fixed cleanly with a template-specialized  functor. Just make the signature something like template<typename T> class writeAttrib {writeAttrib (ALM*, T); ALM* operator() {}}; where the non-specialized version writes like we have it now, and then specialize it for float (maybe also double, but I think we only use floats) to do your mods.

You should also be using %.1f, and not specify a width before the decimal, or you will be padding with spaces -- the client handles the display padding.

Patch-set: 1