Update patch set 1
Patch Set 1:

Oh, you're right. For some reason I thought we had different types. I was hoping to avoid this, but it seems pretty clear that SOE manually added members to all ALMs, instead of doing it generically like we do (in this case). That means to fix this, the special-case components need to be broken out into separate classes and the fillALM method needs to be overridden. You could probably get away with making a ShipComponent or something like that -- I think they all have the same attributes.

Patch-set: 1