Update patch set 1
Patch Set 1:

precision is almost always 0 or 1, but the real problem is that this is JTL and none of the script work for JTL has been done. The mass precision in the chassis scripts is actually set to 0 because it hasn't been touched yet. We are going to need ship components broken out for JTL anyway, so it seems a little temporary to try to fix it in that loop in the generic component. I'm not even sure that component should concretely define the fillALM method and instead should be a virtual stub.

Patch-set: 1