Update patch set 1
Patch Set 1:

We might be able to get away with not specializing ship components, but I'm not sure because JTL crafting seems to be different than normal crafting in that the only "component" class is actually the chassis. This is a bit confusing to me, so I'm not sure exactly how to proceed anymore. In any case though, you'll notice that every component does something a bit different, and most of them are already broken out into specific classes to account for these difference.

Patch-set: 1