Update patch set 1
Patch Set 1:

They are broken out in your version of the source. Component is the base type, so that factory can generically refer to any component. Look in the subdirectories of tangible/component/

And JTL is different in that things like shipweapons share a GOT with the subcomponents to make the shipweapons. Contrast that to ground weapons, whose components are of type meleeweaponcomponent, etc. This means that JTL crafting used the template inheritance and ground crafting used GOT.

Patch-set: 1