Update patch set 1
Patch Set 1:

It is important to note that in the objectFactory in ObjectManager, all the ship components are simply listed with type Component as placeholders -- They are all different GOT's and will have different objects (that probably don't even inherit Component) once they are implemented. If you do a bit of research, you'll probably find that chassis should be it's own SCNO and not inherit Component at all.

Patch-set: 1